+++ CHANGELOG.md
@@ -6,6 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- pre & post build tasks to get and embed the git commit # 
+- info option to display tool and runtime information 
+- roadmap to README.md
+
+### Changed
+
+- extracted command line options definition and parsing to CommandLineOptions class
+- colorized help output
+
 ## v1.1.0 - 2019-04-05
 
 ### Added